+++ packages/datadog_flutter_plugin/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+* Support selective injection of trace contexts with TraceContextInjection configuration item.
 * Bump minimum Dart version to 3.3.0 (Flutter 3.19.0).
 * Add WASM support by removing references to `dart:html` and `package:js`